+  Bugfixes:
@@ -1,10 +1,24 @@
 adns (0.10) unstable; urgency=low

-  * Add list of tested platforms in INSTALL file.
+  Bugfixes:
+  * Actually compile shared libraries by default !
+
+  Portability fixes:
   * Fix <sys/fcntl.h> in client/adnsresfilter.c to <fcntl.h>.
-  * Kill bogus warning, adh-main.c: `arg2' might be used uninitialized ...
   * #include <sys/types.h> and <unistd.h> for fcntl in */*.c.
+  * Cope with compilers that don't do `inline'.
+  * Do not use variadic macro, use stdarg instead (adnslogres.c).
+  * #include "config.h" before "adns.h".
+  * adnstest: setvbuf(stdout,...) before we do first output.
+  * Cast a sizeof(...) in src/event.c to unsigned long before printing.
+  * #include <sys/select.h> if it exists.
+
+  Portability cleanups:
+  * Add list of tested platforms in INSTALL file.
+  * Kill bogus warning, adh-main.c: `arg2' might be used uninitialized ...
   * Add extra {...} near adnslogres.c:167 to kill spurious warning.
+  * Use `printf' instead of `echo -n'.
+

  --

+++ src/event.c
@@ -395,7 +395,8 @@ int adns_processreadable(adns_state ads, int fd, const struct timeval *now) {
       }
       if (udpaddrlen != sizeof(udpaddr)) {
 	adns__diag(ads,-1,0,"datagram received with wrong address length %d"
-		   " (expected %d)", udpaddrlen,sizeof(udpaddr));
+		   " (expected %lu)", udpaddrlen,
+		   (unsigned long)sizeof(udpaddr));
 	continue;
       }
       if (udpaddr.sin_family != AF_INET) {
